+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Remove editing of user name and email from Community Creation flow (fixes #579, #582)
 
 ## [2.1.9] - 2020-10-23
 ## Fixed